requirements.txt
aws-cdk-lib
pyyaml
setuptools
constructs>=10.0.0,<11.0.0